Proxy conf for vip

1. 增加參數, 根據參數決定proxy conf檔
2. logstash 的tag 改成logstash folder 最後一次commit
3. 放log用的mongo db設定(先commit 未上線, 不影響其他功能)

See merge request !9

+++ deployment/proxy-config/idc1/haproxy-idc1/run.sh
@@ -2,16 +2,28 @@
 
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
+if [ "$#" -lt 1 ] ; then
+  echo "should give environment parameter"
+  echo "ex: ./run.sh default"
+  echo "ex: or ./run.sh vip"
+  exit 1
+fi
+
+ENV=$1
+
 NAME='haproxy-idc1'
 
 docker rm -fv $NAME
-docker run -d --name $NAME\
+cmd="docker run -d --name $NAME\
   --restart always \
   -p 9001:9001 \
   -p 9010:9010 \
   -p 9527:9527 \
-  -v $DIR/haproxy.cfg:/usr/local/etc/haproxy/haproxy.cfg:ro \
-  haproxy:1.6
+  -v $DIR/haproxy.${ENV}.cfg:/usr/local/etc/haproxy/haproxy.cfg:ro \
+  haproxy:1.6"
+ 
+echo $cmd
+exec $cmd
 
 echo "# Starting haproxy on idc1"
 head -n 12 haproxy.cfg
